Move GitStorage class into its own file
This class is getting too big to stay in the main
scraped_page_archive.rb file so move it to its own file.
